整合

+++ weatherBriefingSpider/start.py
@@ -7,8 +7,8 @@ cmd1="scrapy crawl briefing -o ./weatherBriefingSpider/brief.json -s FEED_EXPORT
 cmd2="python ../briefGenerator/trainsition.py"
 cmd3="python ../briefGenerator/connect_mongodb.py"
 
-
-os.remove("./weatherBriefingSpider/brief.json")
+if os.path.exists("./weatherBriefingSpider/brief.json"):
+ os.remove("./weatherBriefingSpider/brief.json")
 os.system(cmd1)
 os.system(cmd2)
 os.system(cmd3)
